requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
google-auth-oauthlib==1.2.0
google-auth==2.28.0
google-api-python-client==2.118.0
anthropic==0.42.0
twilio==9.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
pytest==8.0.0
